// app/modules/settings/components/node-data-store-settings/node-data-store-settings-keys.enum.ts
export enum NodeDataStoreSettingsKeys {
  ENABLE_BYTES_USED_WARNING_KEY = 'enableBytesUsedWarning',
  BYTES_USED_WARNING_THRESHOLD_KEY = 'bytesUsedWarningThreshold',
  PURGE_AUTOMATICALLY_KEY = 'purgeAutomatically',
  PURGE_AGE_KEY = 'purgeAge',
  ARCHIVE_ON_PURGE_KEY = 'archiveOnPurge',
  ARCHIVE_DIRECTORY_KEY = 'archiveDirectory',
}